// src/types/redux/siteType.ts
export interface SiteTypeInfo {
    backgroundColor: string
    backgroundImage: string
    borderColor: string
    categoryColor: string
    colorMode: Boolean
    coverImage: string
    entireColor: string
    fontSize: string
    fontWeight: Boolean
    itemColor: string
    logoImage: string
    mainFont: string
    patternSize: number
    patternTilt: string
    priceColor: string
    restaurantID: string
    secondaryFont: string
    secondaryFontSize: string
}